+++ crates/tui/src/http.rs
@@ -6,8 +6,6 @@ mod tests;
 use crate::message::{HttpMessage, Message, MessageSender};
 use async_trait::async_trait;
 use chrono::{DateTime, TimeDelta, Utc};
-use derive_more::derive::Display;
-use indexmap::IndexMap;
 use itertools::Itertools;
 use reqwest::StatusCode;
 use slumber_core::{
@@ -18,9 +16,8 @@ use slumber_core::{
         RequestError, RequestId, RequestRecord, RequestSeed,
         StoredRequestError, TriggeredRequestError,
     },
-    render::{HttpProvider, Prompt, TemplateContext},
+    render::{HttpProvider, TemplateContext},
 };
-use slumber_template::Value;
 use std::{
     collections::{HashMap, hash_map::Entry},
     fmt::Debug,
@@ -30,7 +27,6 @@ use std::{
 use tokio::sync::oneshot;
 use tokio_util::sync::CancellationToken;
 use tracing::warn;
-use uuid::Uuid;
 
 /// Configuration that defines how to render a request
 #[derive(Debug)]
@@ -149,66 +145,12 @@ impl RequestStore {
             start_time: Utc::now(),
             profile_id,
             recipe_id,
-            prompts: IndexMap::new(), // Collect prompts later, as they come in
             cancel_token,
         };
         match self.requests.entry(id) {
             Entry::Vacant(entry) => entry.insert(state),
             Entry::Occupied(_) => panic!("Request {id} started twice"),
         }
-    }
-
-    /// Add a prompt to a building request. The request will remain in the
-    /// building state.
-    pub fn prompt(
-        &mut self,
-        request_id: RequestId,
-        prompt: Prompt,
-    ) -> &RequestState {
-        self.replace(request_id, |mut state| {
-            if let RequestState::Building { prompts, .. } = &mut state {
-                prompts.insert(PromptId::new(), prompt);
-            } else {
-                warn!(
-                    request = ?state,
-                    "Cannot add prompt: not in building state"
-                );
-            }
-            state
-        })
-    }
-
-    /// Send responses to one or more of this request's prompts. This replies to
-    /// the HTTP engine so it can continue building, but does *not* transition
-    /// out of the building state yet.
-    pub fn submit_form(
-        &mut self,
-        request_id: RequestId,
-        responses: Vec<(PromptId, PromptReply)>,
-    ) -> &RequestState {
-        self.replace(request_id, |mut state| {
-            if let RequestState::Building { prompts, .. } = &mut state {
-                for (id, response) in responses {
-                    // Prompts can only be replied to once. If a prompt is not
-                    // in the map, it was probably already replied to which is
-                    // a UI bug.
-                    let prompt = prompts
-                        .shift_remove(&id)
-                        .unwrap_or_else(|| panic!("Prompt {id} not in map"));
-                    response.reply_to(prompt);
-                }
-                // The prompt list is probably empty now, but not necessarily.
-                // Additional prompts may have been added between the form
-                // being submitted and this function being called. In that case,
-                // we expected an additional submission in the future.
-            } else {
-                warn!(
-                    request = ?state,
-                    "Cannot submit prompts: not in building state",
-                );
-            }
-            state
-        })
     }
 
     /// Mark a request as loading. Return the updated state.
@@ -303,7 +245,6 @@ impl RequestStore {
                 start_time,
                 profile_id,
                 recipe_id,
-                prompts: _, // Drop all prompts
                 cancel_token: Some(cancel_token),
             } => {
                 cancel_token.cancel();
@@ -613,13 +554,6 @@ pub enum RequestState {
         /// therefore can't be cancelled independently.
         #[cfg_attr(test, partial_eq(skip))]
         cancel_token: Option<CancellationToken>,
-        /// Any prompts sent by the HTTP engine to be shown to the user. This
-        /// is empty when the request first starts building, and we'll insert
-        /// as prompts are received by the main loop. The UI will show an input
-        /// form with these visible, and upon form submission these will be
-        /// drained out as the responses are forwarded.
-        #[cfg_attr(test, partial_eq(skip))]
-        prompts: IndexMap<PromptId, Prompt>,
     },
 
     /// User cancelled the request mid-build. We don't have the request here
@@ -976,57 +910,6 @@ impl From<&RequestState> for RequestStateSummary {
                 start_time: error.start_time,
                 end_time: error.end_time,
             },
-        }
-    }
-}
-
-/// Unique ID for a single prompt from a request
-///
-/// This is used to correlate a prompt in the request store with an input field
-/// in the UI.
-#[derive(Copy, Clone, Debug, Display, Eq, Hash, PartialEq)]
-pub struct PromptId(Uuid);
-
-impl PromptId {
-    /// Generate a new unique prompt ID
-    pub fn new() -> Self {
-        Self(Uuid::new_v4())
-    }
-}
-
-impl Default for PromptId {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// A UI-provided reply to a prompt. This is the data passed from the UI back
-/// to the request store, to be forwarded via channel to the HTTP engine.
-#[derive(Debug, PartialEq)]
-pub enum PromptReply {
-    Text(String),
-    Select(Value),
-}
-
-impl PromptReply {
-    /// Send this reply to the given prompt. The reply and prompt should be of
-    /// the same input type, otherwise this will panic.
-    fn reply_to(self, prompt: Prompt) {
-        match (prompt, self) {
-            (Prompt::Text { channel, .. }, Self::Text(text)) => {
-                channel.reply(text);
-            }
-            (Prompt::Select { channel, .. }, Self::Select(value)) => {
-                channel.reply(value);
-            }
-            // Prompt/response mismatch. Bug in the prompt form
-            (prompt @ Prompt::Text { .. }, response @ Self::Select(_))
-            | (prompt @ Prompt::Select { .. }, response @ Self::Text(_)) => {
-                panic!(
-                    "Incorrect prompt response type {response:?} \
-                    for prompt {prompt:?}"
-                );
-            }
         }
     }
 }