+++ crates/tui/src/view/common/select.rs
@@ -128,6 +128,7 @@ impl<Item: 'static, State> SelectBuilder<Item, State> {
     }
 
     /// Set the index that should be initially selected
+    #[cfg(test)] // Only used in tests
     pub fn preselect_index(mut self, index: usize) -> Self {
         // If the index is invalid, it will be replaced by 0 in the build()
         self.preselect_index = index;