+++ CHANGELOG.md
@@ -69,7 +69,7 @@ The breaking changes in this release are mostly limited to CLI commands. The onl
 
 - Reconfigured the TUI layout. The main change is that the recipe, profile, and history lists open in an expandable sidebar now
   - Switching between recipes and profiles should feel faster and more intuitive now
-- `prompt()` and `select()` calls are now grouped into a single form pane in the TUI
+- `prompt()` and `select()` calls are now grouped into a single form modal in the TUI
   - Previously, you'd be fed a series of modals one-by-one to fill out
 - Some menu actions have been moved into nested sections for better organization
 - Refactor significant portions of the TUI logic. There should be no user-facing changes, but if you notice any bugs [please report them!](https://github.com/LucasPickering/slumber/issues/new/choose)